added a idle screen so I can better see when it crashes

+++ src/battery.cpp
@@ -15,8 +15,8 @@ void Status::begin(){
 }
 
 void Status::tick(){
-    const uint32_t interval_ms = 5000;
-    static uint32_t start_ms = 0;
+    const uint16_t interval_ms = 5000;
+    static uint16_t start_ms = 0;
 
     if (millis() - start_ms < interval_ms) {
         return;
